+++ .gitignore
@@ -1,4 +1,4 @@
 
 db.sqlite3
 *.pyc
-web_app/__pycache__/*
+auth/__pycache__/*
