+++ authentication/models.py
@@ -1,0 +1,1 @@
+# Place authentication related models here
