+++ authentication/templates/authentication/base.html
@@ -8,7 +8,7 @@
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"
         integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
     {% load static %}
-    <link rel="stylesheet" type="text/css" href="{% static 'web_app/styles.css' %}">
+    <link rel="stylesheet" type="text/css" href="{% static 'auth_styles.css' %}">
     <title>PAS</title>
 </head>
 
@@ -34,10 +34,10 @@
                 <ul class="navbar-nav">
                     {% if not user.is_authenticated %}
                     <li class="nav-item">
-                        <a class="nav-link" href="/login">Login</a>
+                        <a class="nav-link" href="/auth/login">Login</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" href="/register">Register</a>
+                        <a class="nav-link" href="/auth/register">Register</a>
                     </li>
                     {% else %}
                     <li class="nav-item">
@@ -47,7 +47,6 @@
                 </ul>
             </div>
         </nav>
-
         {% endblock nav %}
 
         {% if user.is_authenticated %}
