+++ authentication/templates/authentication/index.html
@@ -1,4 +1,4 @@
-{% extends "web_app/base.html" %}
+{% extends "authentication/base.html" %}
 {% block body %}
 <div class="jumbotron">
     <h1 class="display-4">Policy Assistant - Education Information Management and Surveillance System</h1>
