+++ authentication/templates/authentication/register.html
@@ -1,4 +1,4 @@
-{% extends "web_app/base.html" %}
+{% extends "authentication/base.html" %}
 {% load crispy_forms_tags %}
 {% block body %}
 <form method="post" class="registerForm">
