+++ emis/urls.py
@@ -15,8 +15,11 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path, include
+from authentication.views import index
 
 urlpatterns = [
-    path('', include('web_app.urls')),
+    # Temporary index route
+    path('', index, name="index"),
+    path('auth/', include('authentication.urls')),
     path('admin/', admin.site.urls),
 ]
