--- web_app/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class WebAppConfig(AppConfig):
-    name = 'web_app'
