--- web_app/templates/web_app/foo.html
@@ -1,13 +1,0 @@
-{% extends "web_app/base.html" %}
-{% block body %}
-
-<div class="container grad1">
-    <h3>The number of test scores in the system is: {{ num_test_scores }}</h3>
-    <form action="/testscore/" method="post">
-        {% csrf_token %}
-        {{ form }}
-        <input type="submit" value="Submit">
-    </form>
-</div>
-
-{% endblock body %}
